Raspberry_Pi_setup/requirements_phone.txt
# Requirements for Enhanced Phone-based Gossip System

# Original requirements
openai
openai-whisper
python-dotenv
elevenlabs

# New requirements for Raspberry Pi phone system
gpiozero        # GPIO handling for button detection
pyaudio         # Direct audio recording from microphone
pygame          # Audio playback
RPi.GPIO        # Alternative GPIO library (backup)
lgpio           # Low-level GPIO library for gpiozero
rpi-lgpio       # Alternative lgpio implementation

# Database and file handling (built-in Python modules)
# sqlite3 - built-in
# wave - built-in
# threading - built-in
# uuid - built-in

# Optional but recommended
numpy           # Audio processing support